refactor: narrow cli entrypoint and continuity persistence
