+++ package.json
@@ -19,17 +19,17 @@
   },
   "packageManager": "pnpm@10.11.0",
   "scripts": {
-    "build": "tsc -p tsconfig.build.json",
+    "build": "rimraf dist && tsc -p tsconfig.build.json",
     "clean": "rimraf dist coverage .tmp",
     "dev": "tsx src/cli.ts",
     "lint": "tsc --noEmit -p tsconfig.json",
     "pack:check": "npm pack --dry-run",
     "prepack": "pnpm build",
     "test": "vitest run",
-    "test:cli-smoke": "vitest run test/audit.test.ts test/memory-command.test.ts test/session-command.test.ts",
+    "test:cli-smoke": "vitest run test/audit.test.ts test/memory-command.test.ts test/session-command.test.ts test/wrapper-session-continuity.test.ts",
     "test:dist-cli-smoke": "vitest run test/dist-cli-smoke.test.ts",
     "test:docs-contract": "vitest run test/docs-contract.test.ts",
-    "test:reviewer-smoke": "vitest run test/docs-contract.test.ts test/memory-command.test.ts test/session-command.test.ts test/session-continuity.test.ts",
+    "test:reviewer-smoke": "vitest run test/docs-contract.test.ts test/memory-command.test.ts test/session-command.test.ts test/wrapper-session-continuity.test.ts test/session-continuity.test.ts",
     "test:watch": "vitest",
     "verify:release": "pnpm lint && pnpm test:docs-contract && pnpm test:reviewer-smoke && pnpm test:cli-smoke && pnpm test && pnpm build && pnpm test:dist-cli-smoke && pnpm pack:check"
   },
