+++ src/lib/runtime/runtime-context.ts
@@ -1,15 +1,25 @@
-import { APP_NAME } from "../constants.js";
 import { loadConfig } from "../config/load-config.js";
+import { patchConfigFile } from "../config/write-config.js";
 import { detectProjectContext } from "../domain/project-context.js";
-import { SyncService } from "../domain/sync-service.js";
 import { SessionContinuityStore } from "../domain/session-continuity-store.js";
-import type { AppConfig, LoadedConfig, ProjectContext } from "../types.js";
+import { SyncService } from "../domain/sync-service.js";
+import type {
+  AppConfig,
+  ConfigScope,
+  LoadedConfig,
+  ProjectContext
+} from "../types.js";
 
 export interface RuntimeContext {
   project: ProjectContext;
   loadedConfig: LoadedConfig;
   syncService: SyncService;
   sessionContinuityStore: SessionContinuityStore;
+}
+
+export interface ReloadedRuntimeContext {
+  runtime: RuntimeContext;
+  configUpdatePath: string;
 }
 
 export async function buildRuntimeContext(
@@ -21,6 +31,7 @@ export async function buildRuntimeContext(
   const syncService = new SyncService(project, loadedConfig.config);
   const sessionContinuityStore = new SessionContinuityStore(project, loadedConfig.config);
   await syncService.memoryStore.ensureLayout();
+
   return {
     project,
     loadedConfig,
@@ -29,6 +40,20 @@ export async function buildRuntimeContext(
   };
 }
 
-export function formatWarnings(warnings: string[]): string[] {
-  return warnings.map((warning) => `${APP_NAME} warning: ${warning}`);
+export async function patchConfigAndReloadRuntime(
+  cwd: string,
+  configScope: ConfigScope,
+  updates: Partial<AppConfig>
+): Promise<ReloadedRuntimeContext> {
+  const initialRuntime = await buildRuntimeContext(cwd);
+  const configUpdatePath = await patchConfigFile(
+    initialRuntime.project.projectRoot,
+    configScope,
+    updates
+  );
+
+  return {
+    runtime: await buildRuntimeContext(cwd),
+    configUpdatePath
+  };
 }
